requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==13.0
aiofiles==23.2.1
python-multipart==0.0.9
yfinance==0.2.61
requests==2.31.0
groq==0.8.0
google-generativeai==0.5.4
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
python-dotenv==1.0.1
pydantic==2.7.1